Fix: move the `vela status --ui` command out and add it as a new command `vela top`
